「とりあえず動作する」状態にするためにインポート等を調整する
